@@ -8,10 +8,10 @@
     "homepage": "https://www.loremipsum.at",
     "license": "MIT",
     "require": {
-        "php": ">=7.1",
+        "php": ">=8.1",
         "ext-json": "*",
-        "symfony/http-foundation": "^4.0 || ^5.0 || ^6.0 || ^7.0",
-        "symfony/framework-bundle": "^4.0 || ^5.0 || ^6.0 || ^7.0",
+        "symfony/http-foundation": "^6.0 || ^7.0",
+        "symfony/framework-bundle": "^6.0 || ^7.0",
         "loremipsum/route-generator-bundle": "^0.4.0",
         "twig/twig": "~1.34|~2.4|~3.21"
     },
